Added types to StartScreen
